tests/wpt/web-platform-tests/app-history/navigate-event/transitionWhile-appHistory-back.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    appHistory.navigate("#foo").committed.then(() => {
      assert_true(appHistory.canGoBack);
      appHistory.onnavigate = t.step_func(e => e.transitionWhile(Promise.resolve()));
      appHistory.back().committed.then(t.step_func_done(() => {
        assert_equals(appHistory.entries().length, 2);
        assert_equals(appHistory.current, appHistory.entries()[0]);
      }));
    });
  }, 0);
}, "AppHistoryNavigateEvent's transitionWhile() can intercept appHistory.back()");
</script>
